Final Commit :  Updated testcases with logger and added comparator an…

+++ Framework/src/test/java/com/ndtv/pojo/Main.java
@@ -5,8 +5,8 @@ public class Main
 {
 	private Double temp;
 	private Float feels_like;
-	private Float temp_min;
-	private Float temp_max;
+	private Double temp_min;
+	private Double temp_max;
 	private int pressure;
 	private Double humidity;
 	private int sea_level;
@@ -23,16 +23,16 @@ public class Main
 	public void setFeels_like(Float feels_like) {
 		this.feels_like = feels_like;
 	}
-	public Float getTemp_min() {
+	public Double getTemp_min() {
 		return temp_min;
 	}
-	public void setTemp_min(Float temp_min) {
+	public void setTemp_min(Double temp_min) {
 		this.temp_min = temp_min;
 	}
-	public Float getTemp_max() {
+	public Double getTemp_max() {
 		return temp_max;
 	}
-	public void setTemp_max(Float temp_max) {
+	public void setTemp_max(Double temp_max) {
 		this.temp_max = temp_max;
 	}
 	public int getPressure() {
